requirements.txt
streamlit
torch
numpy
scikit-learn
Pillow
